🐛 Download 'yq' when it is not installed
## Changes Description

Before this commit we simply assumed that the user running the CLI would
have all the tools required by it, installed in their system.

This commit improves the `README.md` file detailing the dependencies of
the project, and it also try to download one of these dependencies when
it is not installed.

## References

- Fix GitHub issue #1;